The change is patched from zhichongh's google3 cl/212453345. Add showing the current point and account name. Add the setting picture to the background of the setting button. Demo video:https://drive.google.com/file/d/1iOmmPRUwl6qtQq0W7xJvZJkSFbnxhXpO/view?usp=sharing
Fixed start CameraActivitiy sometimes black screen bug.
Fixed get one more number of unreviewed video bug.
Sort out CameraActivity code to make reader understand.

+++ app/src/main/java/com/google/android/apps/signalong/utils/FileUtils.java
@@ -41,7 +41,7 @@ public class FileUtils {
 
   public static void clearFile(String filePath) {
     File file = new File(filePath);
-    if (file.exists() && file.isFile()) {
+    if (!TextUtils.isEmpty(filePath) && file.exists() && file.isFile()) {
       file.delete();
     }
   }
